platform/coreapi/requirements/requirements-dev.txt
-r requirements-base.txt

django-debug-toolbar

